Fixed content amounts being cast to int. Replaced Meal plan page icons with Font Awesome icons. Added on hover and title effects to recipe details modal window triggers for more intuitive UI. Made admin editing modal windows larger as well as some text input elements in them. Swaped allergies and password forms in the Account page.

+++ static/scripts/recipes.js
@@ -43,7 +43,7 @@ form_set_edit = function(form_id, id, data=null) {
                 id: parseInt(c["ingredient"]["id"]),
                 name: c["ingredient"]["name"],
             };
-            var amount = parseInt(c["amount"]);
+            var amount = parseFloat(c["amount"]);
             var units = c["units"];
 
             content_add(form_contents, ingredient, amount, units);
